requirements.txt
kafka-python
six
requests
psycopg2-binary